update newsfeed, fix register

+++ src/modules/post/components/PostCardMenu/PostCardMenu.tsx
@@ -5,16 +5,19 @@ import { BiEdit, BiTrash } from "react-icons/bi";
 import { CircularProgress } from "@mui/material";
 import { useParams } from "react-router-dom";
 import MoreVertIcon from "@mui/icons-material/MoreVert";
-import { useTasks } from "../../../../lib/provider/TasksProvider";
 import PostData from "../../interface/post-data";
 import { usePosts } from "../../../../lib/provider/PostsProvider";
+import { useConfirmDialog } from "../../../../lib/provider/ConfirmDialogProvider";
 
 const PostCardMenu = ({ post }: { post: PostData }) => {
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
 
+  const { showConfirmDialog } = useConfirmDialog();
+
   const { roomId } = useParams();
   const { deletePost, deletingPost, setCurrentPost } = usePosts();
+
   return (
     <Box>
       <IconButton
@@ -43,6 +46,7 @@ const PostCardMenu = ({ post }: { post: PostData }) => {
           style={{ display: "flex", padding: 8 }}
           onClick={() => {
             setCurrentPost({ ...post });
+            setAnchorEl(null);
           }}
         >
           <ListItemIcon>
@@ -58,17 +62,22 @@ const PostCardMenu = ({ post }: { post: PostData }) => {
         ) : (
           <MenuItem
             style={{ display: "flex", padding: 8 }}
-            onClick={async () => {
-              await deletePost({
-                room_id: roomId ? roomId : "",
-                id: post.id,
-              });
-              setAnchorEl(null);
-            }}
+            onClick={() =>
+              showConfirmDialog({
+                title: "Xóa bài đăng",
+                content: "Bạn có thực sự muốn xóa bài đăng này không?",
+                onConfirm: async () => {
+                  await deletePost({
+                    room_id: roomId ? roomId : "",
+                    id: post.id,
+                  });
+                  setAnchorEl(null);
+                },
+              })
+            }
           >
             <ListItemIcon>
-              {" "}
-              <BiTrash fontSize="large" />{" "}
+              <BiTrash fontSize="large" />
             </ListItemIcon>
             <Typography variant="inherit" noWrap width="18ch">
               Xóa bài đăng
